+++ src/canalyzerfixes/alCodeFix.ts
@@ -1,0 +1,14 @@
+import * as vscode from 'vscode';
+
+export class ALCodeFix {
+    diagnosticCode: string;
+
+    constructor() {
+        this.diagnosticCode = "";
+    }
+
+    createFix(document: vscode.TextDocument, diagnostic: vscode.Diagnostic): vscode.CodeAction {
+        return new vscode.CodeAction('');
+    }
+
+}