+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.0.27
+ - New functionality added by Stefano Primo:
+   - code actions fixing CodeCop diagnostics 
+     - AA0008 (add parentheses),
+     - AA0137 (remove variable),
+     - AA0139 (add CopyStr)
+   - command to add missing application areas
+
 ## 1.0.26
  - Github issue #74 - Missing global variables in the symbols tree